Extract theme context
