+++ src/main.jsx
@@ -2,15 +2,15 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import CustomThemeProvider from "./theme/ThemeContext";
+import AppThemeProvider from "./theme/ThemeProvider";
 import { BrowserRouter } from "react-router-dom";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
     <React.StrictMode>
-        <CustomThemeProvider>
+        <AppThemeProvider>
             <BrowserRouter>
                 <App />
             </BrowserRouter>
-        </CustomThemeProvider>
+        </AppThemeProvider>
     </React.StrictMode>
 );
