+++ src/sections/Skills.jsx
@@ -10,7 +10,7 @@ import {
     Chip,
     Tooltip,
 } from "@mui/material";
-import { motion, useInView } from "framer-motion";
+import { motion as Motion, useInView } from "framer-motion";
 import SkillIcon from "../components/SkillIcon";
 
 const categoryIcons = {
@@ -40,7 +40,7 @@ export default function Skills() {
 
     return (
         <Box ref={ref} sx={{ px: { xs: 2, md: 4 }, mt: 6 }}>
-            <motion.div
+            <Motion.div
                 initial={{ opacity: 0, y: 40 }}
                 animate={isInView ? { opacity: 1, y: 0 } : {}}
                 transition={{ duration: 0.6 }}
@@ -114,7 +114,7 @@ export default function Skills() {
                         </Card>
                     ))}
                 </Box>
-            </motion.div>
+            </Motion.div>
         </Box>
     );
 }
