+++ src/theme/ThemeContext.js
@@ -1,0 +1,9 @@
+import { createContext, useContext } from "react";
+
+export const ThemeContext = createContext();
+
+export function useThemeContext() {
+    return useContext(ThemeContext);
+}
+
+export default ThemeContext;
